requirements.txt
discord.py[speed]>=2.3.1
git+https://github.com/Rapptz/discord-ext-menus@fbb8803779373357e274e1540b368365fd9d8074
lxml
psutil
asyncpg>=0.12.0
typing_extensions>=4,<5
click
parsedatetime
lru_dict
python-dateutil
pygit2
mmh3
setuptools
aiohttp-jinja2>=1.5
aiohttp-session>=2.12
cryptography